Extract the Degrees to HMS and Degrees to DMS to a seperate helper
